indexer/pyproject.toml
[tool.poetry]
name = "indexer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
pathway = {extras = ["xpack-llm"], version = "^0.15.2"}
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
langchain-openai = "<=0.2.5"
groq = "^0.11.0"
htmltabletomd = "^1.0.0"
unstructured-client = "<0.27.0"
anthropic = "0.40.0"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.154.0"
pyjwt = "^2.10.1"
python-multipart = "^0.0.17"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
